+++ spict/R/checkinp.R
@@ -108,10 +108,11 @@
 #'  \item{"inp$stdevfacC"}{ Factors to multiply the observation error standard deviation of each individual catch observation. Can be used if some observations are more uncertain than others. Must be same length as observation vector. Default: 1.}
 #'  \item{"inp$stdevfacI"}{ Factors to multiply the observation error standard deviation of each individual index observation. Can be used if some observations are more uncertain than others. A list with vectors of same length as observation vectors. Default: 1.}
 #'  \item{"inp$stdevfacE"}{ Factors to multiply the observation error standard deviation of each individual effort observation. Can be used if some observations are more uncertain than others. A list with vectors of same length as observation vectors. Default: 1.}
-#'  \item{"inp$mapsdi"}{ Vector of length equal to the number of index series specifying which indices that should use the same sdi. For example: in case of 3 index series use inp$mapsdi <- c(1, 1, 2) to have series 1 and 2 share sdi and have a separate sdi for series 3. Default: 1:nindex, where nindex is number of index series.}
+#'  \item{"inp$mapsdi"}{ Vector of length equal to the number of index series specifying which indices that should use the same sdi. For example: in case of 3 index series use \code{inp$mapsdi <- c(1, 1, 2)} to have series 1 and 2 share sdi and have a separate sdi for series 3. Default: 1:nindex, where nindex is number of index series.}
 #'  \item{"inp$seasontype"}{ If set to 1 use the spline-based representation of seasonality. If set to 2 use the oscillatory SDE system (this is more unstable and difficult to fit, but also more flexible).}
-#'  \item{"inp$sim.random.effects"}{} ## HERE:
-#'  \item{"inp$sim.priors"}{}         ## HERE:
+#'  \item{"inp$sim.random.effects"}{Should random effects (logB, logF, etc.) be simulated (default) or the same random effects be used (as specified in \code{inp$ini} or in an fitted spict object)?}
+#'  \item{"inp$sim.priors"}{Should any priors be simulated or disregarded (default)? Note, that so far it is only possible to simulate the three default priors for \code{logn}, \code{logalpha}, and \code{logbeta}.}
+#'  \item{"inp$sim.fit"}{Should the estimated parameters from the last fit of a fitted spict object be used for simulation (\code{env$last.par}, default) or the inital values (specified in \code{inp$ini})?. Note, that this only works if a fitted spict object is provided as input to \code{sim.spict}.}
 #' }
 #'
 #' @return An updated list of input variables checked for consistency and with defaults added.
@@ -1195,12 +1196,15 @@ check.inp <- function(inp, verbose = TRUE, mancheck = TRUE){
     checkandadd("manfacs", list(cfac = NULL, ffac = NULL, bfac = NULL), "list")
 
     ## Simulate random effects?
-    if(!"sim.random.effects" %in% names(inp)) inp$sim.random.effects <- 1
+    if(!"sim.random.effects" %in% names(inp)) inp$sim.random.effects <- TRUE
 
     ## Noise ratios (for simulation of priors as random effects)
-    if(!"sim.priors" %in% names(inp)) inp$sim.priors <- 0
+    if(!"sim.priors" %in% names(inp)) inp$sim.priors <- FALSE
     if(!"logalpha" %in% names(inp$ini)) inp$ini$logalpha <- log(1)
     if(!"logbeta" %in% names(inp$ini)) inp$ini$logbeta <- log(1)
+
+    ## Simulate using fitted object (env$last.par)?
+    if(!"sim.fit" %in% names(inp)) inp$sim.fit <- TRUE
 
     # Reorder parameter list
     inp$parlist <- list(logm=inp$ini$logm,